pyproject.toml
[tool.poetry]
name = "ners-verify"
version = "0.1.0"
description = "Backend service for the Non-comercial Expenses Report System"
authors = ["Yoelmy Hernandez <dryoe90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
sqlalchemy = "^2.0.27"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
httpx = "^0.26.0"
pendulum = "^3.0.0"
tenacity = "^8.2.3"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.2.1"
factory-boy = "^3.3.0"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
